move extra includes to includes directory
